fix(lint): clear the clippy backlog, and a token test that failed 1 run in 16
`cargo clippy --all-targets -- -D warnings` reported 17 findings on default
features and 24 with `viewer-distribution`. None were correctness bugs; they
accumulated because nothing has ever linted this repo (issue #3) while the
lints themselves grew underneath it.

Twelve were `collapsible_if` on `if let Some(x) = … { if cond { … } }`. That
shape is only collapsible via let-chains, so clippy flags it **only in edition
2024** — verified against a scratch crate: identical code, flagged at 2024,
ignored at 2021. This crate has been edition 2024 since its initial commit, so
these became lintable through toolchain evolution with no commit touching the
code. bilbycast-edge and bilbycast-manager absorbed the same churn as it
arrived because they are gated; this repo had nowhere for it to surface.

Fixed outright: 13 collapsible_if, field_reassign_with_default,
needless_lifetimes, ptr_arg (`&PathBuf` → `&Path`, which also took
`rotate_file` with it), unnecessary_cast (str0m 0.23 made `numer()` u64, so the
`as u64` became a self-cast), manual_is_multiple_of, question_mark and
while_let_loop.

Allowed with a documented reason in a new `[lints.clippy]`, mirroring the
policy blocks in edge and manager: `new_without_default` (would add `Default`
impls nothing constructs — dead code, which the project bar forbids; edge and
manager both allow it for exactly this), `manual_checked_ops` (the `denom == 0`
guard states "the RTP clock rate is not known yet", which is what a reader of
the WHIP ingest path cares about, not "this division might trap" — edge's
wording), and `result_large_err` (fires on the axum shape
`Result<(), Response>` where the `Err` *is* the HTTP response; boxing it buys
an allocation and an unbox per call site to satisfy a heuristic aimed at
ordinary error enums).

Separately, `tampered_signature_rejected` was failing about one run in sixteen.
It popped the last signature character and pushed a fixed `'0'` — so whenever
the minted signature already ended in `'0'` the "tamper" was a no-op and the
token verified, failing the assertion. `exp` moves every second, so this is a
per-second coin flip; computing the HMAC across 4096 consecutive `exp` values
puts it at 6.35%, against the 6.25% a uniform last nibble predicts. It now
flips the nibble to a value it is not already. Nothing caught it because the
repo has no CI *and* this module only compiles under `viewer-distribution`,
which is off by default.

Verified: clippy clean under default, `viewer-distribution` and
`viewer-distribution-vendored`; 47 tests pass on default features and 122 with
`viewer-distribution`; the fixed test passed 40 consecutive runs.

+++ src/distribution/whip_ingest.rs
@@ -166,7 +166,7 @@ pub(crate) async fn republish_from_session(
                 let is_audio = session.audio_mid == Some(mid);
                 if is_video {
                     // str0m video MediaTime is already the 90 kHz clock.
-                    let pts_90k = rtp_time.numer() as u64;
+                    let pts_90k = rtp_time.numer();
                     asm.push(hub, pts_90k, &data);
                 } else if is_audio {
                     // Opus 48 kHz clock → 90 kHz.
